requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.39.0
frozenlist==1.3.3
idna==3.4
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tokenize-rt==5.0.0
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
jsonlines==4.0.0
streamlit
streamlit-chat
openai
requests
Image
langchain